OU blog: Prohibited users can still comment if public allowed #4046
On personal blog, if there is a post which allows the public to
comment, then all logged-in users can also comment even if they
do not have the 'contributepersonal' capability that is normally
required. This change makes it so the capability is still required
in that case.

If these users wish to comment despite not having the capability
then they can always log out. However, once logged out, their
comments will be subject to the moderation system and therefore
are the responsibility of the person whose blog it is (who we
presumably haven't banned from commenting).