Sort comments by timeposted AND id to avoid random behat failures
Without this change, the order comments are displayed in can be random
if they're posted at the exact same second (as generally happens in
behat testing). This causes a problem for the behat test when it just
goes to delete the second comment on the page, assuming they're in
order.

This likely doesn't affect mysql, but definitely does affect pgsql.